@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from '@angular/common/http';
-import 'rxjs/add/operator/finally';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+import { finalize } from 'rxjs/operators';
 import { HttpObserverService } from './http-observer.service';
 
 @Injectable()
@@ -13,6 +13,6 @@ export class HttpObserverInterceptor implements HttpInterceptor {
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     const timestamp = new Date().getTime();
     this.httpObserverService.addRequest(timestamp, req.url);
-    return next.handle(req).finally(() => this.httpObserverService.addResponse(timestamp, req.url));
+    return next.handle(req).pipe(finalize(() => this.httpObserverService.addResponse(timestamp, req.url)));
   }
 }
